Route select moved to separate component
