+++ src/components/RouteSelector/index.js
@@ -1,0 +1,3 @@
+import { RouteSelector } from './RouteSelector';
+
+export default RouteSelector;
