+++ src/pages/Order/index.js
@@ -1,3 +1,3 @@
-import Order from './Order';
+import { Order } from './Order';
 
 export default Order;
